requirements.txt
altgraph==0.17.4
numpy==2.2.4
opencv-python==4.11.0.86
packaging==24.2
pefile==2023.2.7
pillow==11.1.0
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
pywin32-ctypes==0.2.3
pyzbar==0.1.9